refactor(quiz): integrate quiz phase logic directly into ChoicesSection
- Removed the useQuizPhase hook and integrated its logic into ChoicesSection for improved clarity and maintainability.
- Added state management for selected choice ID and display phase directly within ChoicesSection.
- Updated the import statements to reflect the removal of the useQuizPhase hook and added necessary utility functions for choice correctness checks.
